Add versionId to all resources
The purpose here is to trigger HAPI FHIR's support for FHIR's ETag spec
using weak ETag's for the version id. Note that this algorithm could
result in the FHIR API reporting a changed ETag where the REST API does
not. This is because in the FHIR API the ETag / versionId is derived
entirely from the lastUpdated timestamp. Note that we still do not
support the VREAD operation.